refactor(palette): move items from segment to palette folder
