+++ assets/scripts/palette/SegmentForPalette.scss
